revert unrisk int

+++ features/transaction/buyerorder/src/main/java/com/tokopedia/buyerorder/detail/data/EntityPessenger.java
@@ -74,7 +74,7 @@ public class EntityPessenger {
         this.name = name;
     }
 
-    public long getProductId() {
+    public int getProductId() {
         return productId;
     }
